requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
geopy==2.4.1
Flask-CORS==4.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1
psycopg2-binary==2.9.9  # PostgreSQL adapter for Python 